+++ sdk/storage/test/page_blob_client_test.cpp
@@ -53,10 +53,10 @@ namespace Azure { namespace Storage { namespace Test {
     pageBlobClient.Create(0, m_blobUploadOptions);
 
     EXPECT_EQ(pageBlobClient.GetProperties().ContentLength, 0);
-    pageBlobClient.Resize(static_cast<std::size_t>(2_KB));
-    EXPECT_EQ(pageBlobClient.GetProperties().ContentLength, 2_KB);
-    pageBlobClient.Resize(static_cast<std::size_t>(1_KB));
-    EXPECT_EQ(pageBlobClient.GetProperties().ContentLength, 1_KB);
+    pageBlobClient.Resize(static_cast<int64_t>(2_KB));
+    EXPECT_EQ(static_cast<uint64_t>(pageBlobClient.GetProperties().ContentLength), 2_KB);
+    pageBlobClient.Resize(static_cast<int64_t>(1_KB));
+    EXPECT_EQ(static_cast<uint64_t>(pageBlobClient.GetProperties().ContentLength), 1_KB);
   }
 
   TEST_F(PageBlobClientTest, UploadClear)
@@ -86,8 +86,8 @@ namespace Azure { namespace Storage { namespace Test {
     auto pageRanges = pageBlobClient.GetPageRanges();
     EXPECT_TRUE(pageRanges.ClearRanges.empty());
     ASSERT_FALSE(pageRanges.PageRanges.empty());
-    EXPECT_EQ(pageRanges.PageRanges[0].Offset, 3_KB);
-    EXPECT_EQ(pageRanges.PageRanges[0].Length, 3_KB);
+    EXPECT_EQ(static_cast<uint64_t>(pageRanges.PageRanges[0].Offset), 3_KB);
+    EXPECT_EQ(static_cast<uint64_t>(pageRanges.PageRanges[0].Length), 3_KB);
 
     Azure::Storage::Blobs::GetPageRangesOptions options;
     options.Offset = 4_KB;
@@ -95,8 +95,8 @@ namespace Azure { namespace Storage { namespace Test {
     pageRanges = pageBlobClient.GetPageRanges(options);
     EXPECT_TRUE(pageRanges.ClearRanges.empty());
     ASSERT_FALSE(pageRanges.PageRanges.empty());
-    EXPECT_EQ(pageRanges.PageRanges[0].Offset, 4_KB);
-    EXPECT_EQ(pageRanges.PageRanges[0].Length, 1_KB);
+    EXPECT_EQ(static_cast<uint64_t>(pageRanges.PageRanges[0].Offset), 4_KB);
+    EXPECT_EQ(static_cast<uint64_t>(pageRanges.PageRanges[0].Length), 1_KB);
 
     auto snapshot = pageBlobClient.CreateSnapshot().Snapshot;
     // |_|_|_|x|  |x|x|_|_| This is what's in snapshot
@@ -112,9 +112,9 @@ namespace Azure { namespace Storage { namespace Test {
     ASSERT_FALSE(pageRanges.ClearRanges.empty());
     ASSERT_FALSE(pageRanges.PageRanges.empty());
     EXPECT_EQ(pageRanges.PageRanges[0].Offset, 0);
-    EXPECT_EQ(pageRanges.PageRanges[0].Length, 1_KB);
-    EXPECT_EQ(pageRanges.ClearRanges[0].Offset, 3_KB);
-    EXPECT_EQ(pageRanges.ClearRanges[0].Length, 1_KB);
+    EXPECT_EQ(static_cast<uint64_t>(pageRanges.PageRanges[0].Length), 1_KB);
+    EXPECT_EQ(static_cast<uint64_t>(pageRanges.ClearRanges[0].Offset), 3_KB);
+    EXPECT_EQ(static_cast<uint64_t>(pageRanges.ClearRanges[0].Length), 1_KB);
   }
 
   TEST_F(PageBlobClientTest, UploadFromUri)
